moving announcments filter to announcements from partnerCards
