+++ edsdme/components/PartnerCards.js
@@ -5,22 +5,6 @@ import './SinglePartnerCard.js';
 const miloLibs = getLibs();
 const { html, LitElement, css, repeat } = await import(`${miloLibs}/deps/lit-all.min.js`);
 
-export function filterRestrictedCardsByCurrentSite(cards) {
-  const currentSite = window.location.pathname.split('/')[1];
-  return cards.filter((card) => {
-    const cardUrl = card?.contentArea?.url;
-    if (!cardUrl) return false;
-    try {
-      const cardSite = new URL(cardUrl).pathname.split('/')[1];
-      return currentSite === cardSite;
-    } catch (error) {
-      // eslint-disable-next-line no-console
-      console.error(`Invalid URL: ${cardUrl}`, error);
-      return false;
-    }
-  });
-}
-
 export default class PartnerCards extends LitElement {
   static styles = [
     partnerCardsStyles,
@@ -239,8 +223,6 @@ export default class PartnerCards extends LitElement {
         if (window.location.hostname === 'partners.adobe.com') {
           apiData.cards = apiData.cards.filter((card) => !card.contentArea.url?.includes('/drafts/'));
         }
-        // Filter announcements by current site
-        apiData.cards = filterRestrictedCardsByCurrentSite(apiData.cards);
         // eslint-disable-next-line no-return-assign
         apiData.cards.forEach((card, index) => card.orderNum = index + 1);
         this.allCards = apiData.cards;
